+++ e2e-tests/workspace-compat/CHANGELOG.md
@@ -1,5 +1,7 @@
 # workspace-compat-e2e-test
 
+## 0.0.49
+
 ## 0.0.48
 
 ## 0.0.47